app/gmail_client.py
"""
Gmail API client wrapper.
"""
import base64
import os
import pickle
from email.mime.text import MIMEText
from pathlib import Path
from typing import Dict, List, Optional

from google.auth.transport.requests import Request
from google_auth_oauthlib.flow import InstalledAppFlow
from googleapiclient.discovery import build

from app.settings import settings


class GmailClient:
    """Gmail API client wrapper."""
    
    SCOPES = ['https://www.googleapis.com/auth/gmail.readonly', 
              'https://www.googleapis.com/auth/gmail.send']
    
    def __init__(self, credentials_path: Optional[Path] = None):
        """Initialize Gmail client with credentials."""
        # === Task-05 =============================
        # Initialize Gmail client and get service
        # Hints: Store credentials path and get Gmail service
        raise NotImplementedError("Student TODO - Initialize Gmail client")
        
    def _get_gmail_service(self):
        """Get Gmail API service."""
        # === Task-06 =============================
        # Get authenticated Gmail service
        # Hints: Handle credentials loading, refresh, and OAuth flow
        raise NotImplementedError("Student TODO - Get Gmail service")
    
    def get_thread(self, thread_id: str) -> Dict:
        """Get a thread by ID with all messages."""
        # === Task-07 =============================
        # Get Gmail thread by ID
        # Hints: Use Gmail API threads().get()
        raise NotImplementedError("Student TODO - Get thread")
    
    def get_inbox_emails(self, max_results: int = 10) -> List[Dict]:
        """Get top emails from the inbox.
        
        Args:
            max_results: Maximum number of emails to return
            
        Returns:
            List of email messages with basic info
        """
        # === Task-08 =============================
        # Get emails from inbox with metadata
        # Hints: Use messages().list() and messages().get()
        raise NotImplementedError("Student TODO - Get inbox emails")
    
    def extract_thread_content(self, thread_data: Dict) -> List[Dict]:
        """Extract the content from a thread's messages."""
        # === Task-09 =============================
        # Extract message content from thread data
        # Hints: Parse headers and decode message body
        raise NotImplementedError("Student TODO - Extract thread content")
    
    def send_reply(self, thread_id: str, to: str, subject: str, body: str) -> Dict:
        """Send a reply to a thread."""
        # === Task-10 =============================
        # Send email reply to thread
        # Hints: Create MIME message and use messages().send()
        raise NotImplementedError("Student TODO - Send reply") 