pyproject.toml
[tool.poetry]
name = "inbox-helper"
version = "0.1.0"
description = "A lightweight agentic email assistant"
authors = ["Your Name <your.email@example.com>"]
readme = "README.md"
packages = [{include = "app"}]

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.110.0"
uvicorn = "^0.28.0"
sqlmodel = "^0.0.14"
pydantic = "^2.5.0"
pydantic-settings = "^2.1.0"
google-api-python-client = "^2.120.0"
google-auth-httplib2 = "^0.2.0"
google-auth-oauthlib = "^1.2.0"
openai = "^1.0.0"
httpx = "^0.27.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
pytest-asyncio = "^0.23.0"
pytest-cov = "^4.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
asyncio_mode = "auto" 